Created constants file for worker scripts
