+++ workers/networkUtils/xhrReqs.js
@@ -6,51 +6,11 @@
 
 // Document imports
 const https = require("https");
-const FormData = require("form-data");
-
-// FormData for POST request
-var formdata = new FormData();
-formdata.append("distribradio", "alldistribs");
-formdata.append("depts", "no_value");
-formdata.append("periods", "no_value");
-formdata.append("distribs", "no_value");
-formdata.append("distribs_i", "no_value");
-formdata.append("distribs_wc", "no_value");
-formdata.append("deliverymodes", "no_value");
-formdata.append("pmode", "public");
-formdata.append("term", "");
-formdata.append("levl", "");
-formdata.append("fys", "n");
-formdata.append("wrt", "n");
-formdata.append("pe", "n");
-formdata.append("review", "n");
-formdata.append("crnl", "no_value");
-formdata.append("classyear", "2008");
-formdata.append("searchtype", "Subject Areas");
-formdata.append("termradio", "selectterms");
-formdata.append("terms", "no_value");
-formdata.append("terms", "202201");
-formdata.append("deliveryradio", "selectdelivery");
-formdata.append("subjectradio", "selectsubjects");
-formdata.append("depts", "COSC");
-formdata.append("hoursradio", "allhours");
-formdata.append("sortorder", "dept");
-
-// Request Options
-const options = {
-  hostname: "oracle-www.dartmouth.edu",
-  port: 443,
-  path: "/dart/groucho/timetable.display_courses",
-  method: "POST",
-  headers: {
-    ...formdata.getHeaders(),
-    "Content-Length": formdata.getLengthSync(),
-  },
-};
+const workerConstants = require("../constants");
 
 function timetablePostRequest() {
   return new Promise(function (resolve, reject) {
-    var req = https.request(options, (res) => {
+    var req = https.request(workerConstants.CS_POST_OPTIONS, (res) => {
       var body = [];
 
       //Log status code
@@ -59,7 +19,6 @@ function timetablePostRequest() {
       // Concatenate Data as it's recieved
       res.on("data", (d) => {
         const htmlString = d.toString("utf-8");
-        console.log(htmlString);
         body.push(htmlString);
       });
 
@@ -78,7 +37,7 @@ function timetablePostRequest() {
     });
 
     // Write form data buffer into POST request
-    req.write(formdata.getBuffer());
+    req.write(workerConstants.CS_FORM_DATA.getBuffer());
 
     // Reject promise on error
     req.on("error", (error) => {
